Add styling to mission page
