+++ src/components/mission.css
@@ -1,0 +1,65 @@
+.mission_table {
+    width: 80%;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 3rem;
+    max-height: 700px;
+    overflow-y: scroll;
+    border-collapse: collapse;
+    width: 100%;
+    border-bottom: 2px solid #d3d3d3;
+}
+
+td,
+th {
+  border: 2px solid black;
+  text-align: left;
+  padding: 8px;
+}
+
+th {
+  vertical-align: top;
+  font-size: 20px;
+}
+
+tr {
+  color: black;
+}
+
+.member_status {
+  background-color: rgba(0, 0, 0, 0.529);
+  border-radius: 7px;
+  padding: 0.5rem;
+  white-space: nowrap;
+  color: #fff;
+  font-weight: bold;
+  font-size: 16px;
+  text-align: center;
+}
+
+.description {
+  font-size: 18px;
+  opacity: 0.7;
+}
+
+.join-botton,
+.leave-button {
+  white-space: nowrap;
+  padding: 0.5rem;
+  font-size: 18px;
+  width: 150px;
+  border-radius: 4px;
+  outline: transparent;
+  border: transparent;
+  cursor: pointer;
+}
+
+.join-button {
+  background: rgb(192, 1, 1);
+  color: #fff;
+}
+
+.leave-button {
+  background: #fff;
+  color: rgb(192, 1, 1);
+}
